heroku branch ready, addMessage route is configured correctly
