+++ package.json
@@ -1,24 +1,18 @@
 {
-  "name": "cib-reserve-back",
-  "version": "1.0.0",
-  "description": "This script is invoked upon request from users application to manipulate availabe time slots for reservation and return the service providing counters and availabe time slots.",
-  "main": "index.js",
+  "name": "functions",
+  "description": "Cloud Functions for Firebase",
+  "scripts": {
+    "serve": "firebase serve --only functions",
+    "shell": "firebase experimental:functions:shell",
+    "start": "npm run shell",
+    "deploy": "firebase deploy --only functions",
+    "logs": "firebase functions:log"
+  },
   "dependencies": {
     "body-parser": "^1.18.2",
-    "express": "^4.16.3"
+    "express": "^4.16.3",
+    "firebase-admin": "^5.12.0",
+    "firebase-functions": "^1.0.1"
   },
-  "devDependencies": {},
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ahmedHusseinF/cib-reserve-back.git"
-  },
-  "author": "Xyntechs, Mariam",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/ahmedHusseinF/cib-reserve-back/issues"
-  },
-  "homepage": "https://github.com/ahmedHusseinF/cib-reserve-back#readme"
+  "private": true
 }
